src/main/webapp/app/entities/sol/update/sol-update.component.spec.ts
import { ComponentFixture, TestBed } from '@angular/core/testing';
import { HttpResponse } from '@angular/common/http';
import { HttpClientTestingModule } from '@angular/common/http/testing';
import { FormBuilder } from '@angular/forms';
import { ActivatedRoute } from '@angular/router';
import { RouterTestingModule } from '@angular/router/testing';
import { of, Subject, from } from 'rxjs';

import { SolService } from '../service/sol.service';
import { ISol, Sol } from '../sol.model';

import { SolUpdateComponent } from './sol-update.component';

describe('Sol Management Update Component', () => {
  let comp: SolUpdateComponent;
  let fixture: ComponentFixture<SolUpdateComponent>;
  let activatedRoute: ActivatedRoute;
  let solService: SolService;

  beforeEach(() => {
    TestBed.configureTestingModule({
      imports: [HttpClientTestingModule, RouterTestingModule.withRoutes([])],
      declarations: [SolUpdateComponent],
      providers: [
        FormBuilder,
        {
          provide: ActivatedRoute,
          useValue: {
            params: from([{}]),
          },
        },
      ],
    })
      .overrideTemplate(SolUpdateComponent, '')
      .compileComponents();

    fixture = TestBed.createComponent(SolUpdateComponent);
    activatedRoute = TestBed.inject(ActivatedRoute);
    solService = TestBed.inject(SolService);

    comp = fixture.componentInstance;
  });

  describe('ngOnInit', () => {
    it('Should update editForm', () => {
      const sol: ISol = { id: 456 };

      activatedRoute.data = of({ sol });
      comp.ngOnInit();

      expect(comp.editForm.value).toEqual(expect.objectContaining(sol));
    });
  });

  describe('save', () => {
    it('Should call update service on save for existing entity', () => {
      // GIVEN
      const saveSubject = new Subject<HttpResponse<Sol>>();
      const sol = { id: 123 };
      jest.spyOn(solService, 'update').mockReturnValue(saveSubject);
      jest.spyOn(comp, 'previousState');
      activatedRoute.data = of({ sol });
      comp.ngOnInit();

      // WHEN
      comp.save();
      expect(comp.isSaving).toEqual(true);
      saveSubject.next(new HttpResponse({ body: sol }));
      saveSubject.complete();

      // THEN
      expect(comp.previousState).toHaveBeenCalled();
      expect(solService.update).toHaveBeenCalledWith(sol);
      expect(comp.isSaving).toEqual(false);
    });

    it('Should call create service on save for new entity', () => {
      // GIVEN
      const saveSubject = new Subject<HttpResponse<Sol>>();
      const sol = new Sol();
      jest.spyOn(solService, 'create').mockReturnValue(saveSubject);
      jest.spyOn(comp, 'previousState');
      activatedRoute.data = of({ sol });
      comp.ngOnInit();

      // WHEN
      comp.save();
      expect(comp.isSaving).toEqual(true);
      saveSubject.next(new HttpResponse({ body: sol }));
      saveSubject.complete();

      // THEN
      expect(solService.create).toHaveBeenCalledWith(sol);
      expect(comp.isSaving).toEqual(false);
      expect(comp.previousState).toHaveBeenCalled();
    });

    it('Should set isSaving to false on error', () => {
      // GIVEN
      const saveSubject = new Subject<HttpResponse<Sol>>();
      const sol = { id: 123 };
      jest.spyOn(solService, 'update').mockReturnValue(saveSubject);
      jest.spyOn(comp, 'previousState');
      activatedRoute.data = of({ sol });
      comp.ngOnInit();

      // WHEN
      comp.save();
      expect(comp.isSaving).toEqual(true);
      saveSubject.error('This is an error!');

      // THEN
      expect(solService.update).toHaveBeenCalledWith(sol);
      expect(comp.isSaving).toEqual(false);
      expect(comp.previousState).not.toHaveBeenCalled();
    });
  });
});
